rag-whatsapp-bot/requirements.txt
flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langchain-ollama==0.3.6
langchain-huggingface==0.3.1
sentence-transformers==5.0.0
faiss-cpu==1.11.0.post1
torch==2.7.0
transformers==4.53.1
numpy==2.2.6
pandas==2.3.1
scikit-learn==1.7.1
requests==2.32.3
gunicorn==21.2.0 